Remove the dict constraint in ForwardRef._eval_type
This improves compatibility with Pyston, which doesn't use a dict for globals.
I'm not sure if there is a better solution, but just removing the check seems ok:
eval() will still do type checking, and though the error messages won't be
as nice I think it's ok because this is all internal (the locals/globals seem
to come from f_locals/f_globals).