get_type_hints(): find the right globalns for classes and modules
This makes the default behavior (without specifying `globalns` manually) more
predictable for users.

Implementation for classes assumes has a `__module__` attribute and that module
is present in `sys.modules`.  It does this recursively for all bases in the
MRO.  For modules, the implementation just uses their `__dict__` directly.

This is backwards compatible, will just raise fewer exceptions in naive user
code.