Modify docs to comment out unpopulated toctrees (#987)
Prior to this, the toctree directives pointed at numerous planned but
unpopulated docs. The result is a mostly empty site with the real
content buried.

To revert without discarding work, the toctrees which list empty files
are commented out, and new tables of contents are added to link to
`libraries` and `stubs` (the existing, populated sections).

Stub docs are left in place, producing sphinx warnings about unlinked
docs. These are expected.

The link to the typing module docs has been restored, now in a
dedicated section.