Document how to test type annotations (#1071)
This is a slightly broader refactor than just testing. It also
consolidates information about checking type coverage/completeness.

This originates from a thread on the mypy tracker [1]. In terms of
presentation, the goal is to present guidance and offer up several
options, many of which were proposed by contributors to that thread.

Several of the goals from that thread were not achieved here,
including documentation covering stubgen and monkeytype, stubtest, and
potentially more. However, the document is written such that it should
be possible to add a section on "Generating Annotations" as was
planned earlier.

[1]: https://github.com/python/mypy/issues/11506