Make pytype conformance test a little faster. (#1559)
This change gets the pytype test down to ~30s. It's still considerably
slower than the others, but any further improvements will likely require
untangling the sad state of affairs that is
https://github.com/google/pytype/issues/1501, which will take some time.

* Runs pytype's 'check' option through pytype.io.check_py to skip the
  costly type inference that it does by default.
* Adds a progress bar so that it doesn't look like pytype is hanging.

This generates a few changes to results/pytype/ due to error messages
being formatted and parsed in a different way.