Modernise more description in stubs distribution (#1577)
A common confusion in the Python ecosystem is the difference between
distribution names and package names. I've seen users be slightly
confused by types-* and *-stubs, so it can't hurt to be more explicit

Remove mentions of ancient pip and distutils

One maybe controversial thing is the loosening of a SHOULD to a MAY.
This is meant to reflect the reality of how e.g. typeshed operates,
which I think is pretty well thought through and I would be opposed to
changing.