Updated conformance test results for pyright 1.1.347 (#1588)
Updated conformance test results for pyright 1.1.347, which fixed a bunch of non-conformance issues. Fixed a couple of bugs in the conformance tests that were found in the process: `protocols_self` and `protocols_recursive` were using the wrong variance (which went uncaught because of a previous bug in pyright's variance inference algorithm).