Fixed a bug in the conformance tests for TypeVarTuples and removed a … (#1593)
Fixed a bug in the conformance tests for TypeVarTuples and removed a note that I had left in previously. This section in the spec is unfortunately worded and can be interpreted in different ways, but I think I now understood what the PEP authors meant. This change to the tests doesn't change any of the scoring for the four type checkers. Conformance with this part of the spec will require fixes in both pyright and mypy.