Added code to make the compliance test more deterministic.  (#1600)
Added code to make the compliance test more deterministic. In particular, pytype doesn't appear to report errors in the same order all the time if there are multiple errors on the same line. This change forces the pytype output to a deterministic ordering. Also deleted details from pytype errors, which can also be nonrepeatable from one run to the next. It also deletes caches for mypy and pyre so reported timings are more repeatable. Also rounded run times to one decimal place.