Conformance tests: Fix automated scoring for dataclasses_usage.toml (#1684)
Conformance tests: Mark error for unannotated "field()" as optional, plus pytype changes

The spec doesn't explicitly mention this case. Since it produces a runtime error,
it is reasonable for type checkers to error on it, but it should not be required.

For pytype, mark a few errors as optionally appearing on a different line.
Also, change "x ** 2" to "x * x"; pytype infers "x ** 2" as being of type "float | int"
for whatever reason.