Updated conformance test results for pyright 1.1.359. (#1719)
Most of the changes are due to improvements in error message wording. Fixed a couple of bugs in the tests that were uncovered by recent changes in pyright. Pyright 1.1.359 fixed several conformance issues (specifically with constructor calls and forward references in PEP 695 syntax) but also introduced two conformance regressions that were caught by the tests.