Manually adjust results
Edit the notes where there were obvious updates, particularly related
to changes because of `assert_type` being supported.

Mark a few tests as Passing that were not before, and some others as
either Unsupported or Partial where they were previously marked as
better-supported (many of these are again related to `assert_type`;
improved support for that makes it clearer that in a few cases Pyre was
not conformant but appeared to be due to not throwing errors).