Fixed a bug in the protocols_generic.py test related to protocol matc… (#1864)
Fixed a bug in the protocols_generic.py test related to protocol matching when method-scoped type variables were used. Both pyright and mypy were previously reported as failing this test, but they were actually working correctly. This was a bug in the original test case (which I wrote).

Updated to the latest version of pytest (2024.10.11). It updates many error messages but has no substantive changes to the conformance results.