Removed support for the pytype type checker (#2010)
Removed support for the pytype type checker.

It is becoming increasingly difficult to maintain support for pytype in the conformance tests. The maintainers of pytype have indicated that they do not have plans to make their type checker compliant with the typing spec, and they've agreed that it makes sense to eliminate it from the conformance test results.