conformance tests: drop timing information
Timing information is nice to have but the current data is not very reliable,
as it is the timing on whatever machine was used by the person who most recently
regenerated the results. It's also a big source of merge conflicts, basically
meaning that every pair of changes to the conformance suite will merge conflict
with each other.

To make it more pleasant to work with the conformance suite, let's just drop the timings.