spec: Rewrite TypedDict spec (#2072)
* spec: Rewrite TypedDict spec

This is an edit of the TypedDict spec for clarity and flow. My goal was to
unify the pieces of the spec that derive from the various PEPs into a coherent
whole. I removed excessive examples and motivations: the spec should specify,
not justify. The length of the spec chapter is reduced by more than half.

This change is on top of #2068 (adding PEP 728).

The general approach I took is to first define the kinds of TypedDicts that
can exist, then explain the syntax for defining TypedDicts, then discuss
other aspects of TypedDict types.

I introduce some new terminology around PEP 728 to make it easier to talk
about the different kinds of TypedDict. TypedDicts are defined to have a
property called openness, which can have three states:
- Open: all TypedDicts prior to PEP 728
- Closed: no extra keys are allowed (closed=True)
- With extra items: extra_items=... from PEP 728

I retained existing text where it made sense but also wrote some from
scratch.

* Apply suggestions from code review

Co-authored-by: Joren Hammudoglu <jhammudoglu@gmail.com>

* feedback

* Carl feedback

* more clarity

---------

Co-authored-by: Joren Hammudoglu <jhammudoglu@gmail.com>